calories_service/requirements.txt
Pillow==8.4.0
torch==1.8.1
torchvision==0.9.1
scikit-learn==0.24.1
efficientnet-pytorch==0.6.3
fastapi
python-multipart
uvicorn